+++ frontend/src/components/FleetView.tsx
@@ -25,7 +25,8 @@ import { useLicense } from '@/context/LicenseContext';
 import { PaidGate } from './PaidGate';
 import FleetSnapshots from './FleetSnapshots';
 import { toast } from '@/components/ui/toast-store';
-import { LabelDot, type Label as StackLabel } from './LabelPill';
+import { LabelDot } from './LabelPill';
+import { type Label as StackLabel } from './label-types';
 import { MultiSelectCombobox } from '@/components/ui/multi-select-combobox';
 import { formatVersion } from '@/lib/version';
 import { CursorProvider, Cursor, CursorFollow, CursorContainer } from '@/components/animate-ui/primitives/animate/cursor';
