+++ frontend/src/components/settings/LabelsSection.tsx
@@ -23,7 +23,8 @@ import { apiFetch } from '@/lib/api';
 import { toast } from '@/components/ui/toast-store';
 import { PaidGate } from '../PaidGate';
 import { CapabilityGate } from '../CapabilityGate';
-import { LabelDot, LABEL_COLORS, MAX_LABELS_PER_NODE, type Label, type LabelColor } from '../LabelPill';
+import { LabelDot } from '../LabelPill';
+import { LABEL_COLORS, MAX_LABELS_PER_NODE, type Label, type LabelColor } from '../label-types';
 
 export function LabelsSection() {
     const [labels, setLabels] = useState<Label[]>([]);
